MOBILSDK-1026: Renamed OnExit sample to Contact for consistency, updated logo, sdk version
