+++ ContactSurvey/app/src/main/AndroidManifest.xml
@@ -1,0 +1,67 @@
+<?xml version="1.0" encoding="utf-8"?>
+<manifest xmlns:android="http://schemas.android.com/apk/res/android"
+    package="com.foresee.contactSurvey"
+    android:versionCode="1"
+    android:versionName="1.0" >
+
+    <uses-sdk android:minSdkVersion="4" android:targetSdkVersion="17"/>
+
+    <!-- ForeSee SDK permissions -->
+    <uses-permission android:name="android.permission.INTERNET"/>
+    <uses-permission android:name="android.permission.ACCESS_NETWORK_STATE"/>
+    <uses-permission android:name="android.permission.ACCESS_WIFI_STATE"/>
+
+
+    <application
+        android:name="com.foresee.contactSurvey.CustomApplication"
+        android:label="@string/app_name"
+        android:allowBackup="false">
+
+        <activity android:name="com.foresee.sdk.tracker.app.SurveyInviteActivity"
+                  android:configChanges="orientation|screenSize|keyboard|keyboardHidden"
+                  android:theme="@android:style/Theme.Translucent.NoTitleBar">
+            <intent-filter>
+                <action android:name="com.fsr.tracker.app.LAUNCH_INVITE"/>
+                <category android:name="android.intent.category.DEFAULT"/>
+            </intent-filter>
+        </activity>
+        <activity android:name="com.foresee.sdk.tracker.app.OnExitSurveyInviteActivity"
+                  android:configChanges="orientation|screenSize|keyboard|keyboardHidden"
+                  android:theme="@android:style/Theme.Translucent.NoTitleBar">
+            <intent-filter>
+                <action android:name="com.fsr.tracker.app.LAUNCH_INVITE"/>
+                <category android:name="android.intent.category.DEFAULT"/>
+            </intent-filter>
+        </activity>
+        <activity android:name="com.foresee.sdk.tracker.app.SurveyActivity"
+                  android:configChanges="orientation|screenSize|keyboard|keyboardHidden">
+            <intent-filter>
+                <action android:name="com.fsr.tracker.app.LAUNCH_SURVEY"/>
+                <category android:name="android.intent.category.DEFAULT"/>
+            </intent-filter>
+        </activity>
+        <!-- ForeSee cxReplay -->
+        <service android:name="com.foresee.mobileReplay.jobQueue.JobQueueIntentService"
+                 android:process=":jobqueue"/>
+        <service android:name="com.foresee.mobileReplay.services.PerfLogService"/>
+        <receiver android:name="com.foresee.mobileReplay.receivers.ConnectivityChangeEventReceiver">
+            <intent-filter>
+                <action android:name="android.net.wifi.STATE_CHANGE"/>
+                <action android:name="android.net.conn.CONNECTIVITY_CHANGE"/>
+            </intent-filter>
+        </receiver>
+        <!-- Required for 'localNotificationDelay' config key -->
+        <receiver android:name="com.foresee.sdk.tracker.services.NotificationReceiver"/>
+
+        <activity
+            android:name="com.foresee.contactSurvey.OnExit"
+            android:label="@string/app_name"
+            android:configChanges="orientation|screenSize|keyboard|keyboardHidden">
+            <intent-filter>
+                <action android:name="android.intent.action.MAIN" />
+                <category android:name="android.intent.category.LAUNCHER" />
+            </intent-filter>
+        </activity>
+    </application>
+</manifest>    
+
