+++ ContactSurvey/app/src/main/res/values/strings.xml
@@ -1,0 +1,13 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+
+    <string name="hello">This is a demo of an "on-exit" survey implementation.</string>
+    <string name="app_name">Contact Survey Demo</string>
+	<string name="selectModeLabel">Contact survey notification mode</string>
+	<string-array name="onExitModes">
+	    <item>Email or SMS notification</item>
+	    <item>Local notification</item>
+	</string-array>
+	    <string name="launch_invite">Launch invitation</string>
+	    <string name="reset_counters">Reset counters</string>
+</resources>
