+++ ContactSurvey/build.gradle
@@ -3,6 +3,7 @@
 buildscript {
     repositories {
         jcenter()
+        mavenCentral()
     }
     dependencies {
         classpath 'com.android.tools.build:gradle:2.2.3'
@@ -15,6 +16,10 @@ buildscript {
 allprojects {
     repositories {
         jcenter()
+        mavenCentral()
+        flatDir {
+            dirs 'libs'
+        }
     }
 }
