--- OnExitSurvey/app/src/main/java/com/foresee/onexitsurvey/CustomApplication.java
@@ -1,13 +1,0 @@
-package com.foresee.onexitsurvey;
-
-import com.foresee.sdk.ForeSee;
-
-import android.app.Application;
-
-public class CustomApplication extends Application {
-	@Override
-	public void onCreate() {
-		// Notify ForeSee SDK of application start
-		ForeSee.start(this);
-	}
-}
