--- OnExitSurvey/app/src/main/res/raw/foresee_configuration.json
@@ -1,30 +1,0 @@
-{
-	"clientId":"Ij6P1lfZHchO/co10lQ4BQ==",
-	"repeatDaysAfterDecline":5,
-	"repeatDaysAfterComplete":5,
-	"repeatDaysAfterAccept":3,
-	"notificationType":"CONTACT",
-	"supportsReinvitation":true,
-	"sessionReplayEnabled":false,
-	"whitelistedHosts":
-	[
-		"www.google.com",
-		"www.hotmail.com"
-	],
-	"measures":
-	[
-		{
-			"surveyId":"app_test_1",
-			"daysSinceLaunch":5,
-			"launchCount":2,
-			"significantEventThresholds":
-            {
-                "app_test_1": 4
-            }
-		}
-	],
-    "cppParameters":
-    {
-        "sample_app":"On Exit Survey 2.0"
-    }
-}
